Changes from the meeting
