--- src/main/java/net/eventsexpress/app/Config/app.properties
@@ -1,2 +1,0 @@
-baseURL = https://eventsexpress-test.azurewebsites.net/landing
-elementTimeOut = 10
