--- src/main/java/net/eventsexpress/app/Utils/Utils.java
@@ -1,12 +1,0 @@
-package net.eventsexpress.app.Utils;
-
-public class Utils {
-
-    public static void sleep(long m) {
-        try {
-            Thread.sleep(m);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
-}
