--- src/main/java/net/eventsexpress/app/WebElements/ButtonElement.java
@@ -1,5 +1,0 @@
-package net.eventsexpress.app.WebElements;
-
-public class ButtonElement {
-
-}
