--- src/main/java/net/eventsexpress/app/WebElements/InputElement.java
@@ -1,5 +1,0 @@
-package net.eventsexpress.app.WebElements;
-
-public class InputElement {
-
-}
