--- src/test/java/net/eventsexpress/app/TestConfig/test.properties
@@ -1,1 +1,0 @@
-driver = edge
