+++ src/test/java/net/eventsexpress/app/Utils/Utils.java
@@ -1,0 +1,12 @@
+package net.eventsexpress.app.Utils;
+
+public class Utils {
+
+    public static void sleep(long m) {
+        try {
+            Thread.sleep(m);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+}
