+++ src/test/java/net/eventsexpress/app/WebElements/InputElement.java
@@ -1,0 +1,5 @@
+package net.eventsexpress.app.WebElements;
+
+public class InputElement {
+
+}
