+++ src/test/java/net/eventsexpress/app/common/BasePage.java
@@ -1,0 +1,52 @@
+package net.eventsexpress.app.common;
+
+public class BasePage extends BaseWrapper {
+
+//    private AdminPanel adminPanel;
+//    private EventDetails eventDetails;
+//    private Footer footer;
+//    private Header header;
+//    private LandingPage landingPage;
+//    private Navigation navigation;
+//
+//    public Privacy privacy;
+//
+//    public BasePage(WebDriver driver) {
+//        super(driver);
+//        adminPanel = new AdminPanel(driver);
+//        eventDetails = new EventDetails(driver);
+//        footer = new Footer(driver);
+//        header = new Header(driver);
+//        landingPage = new LandingPage(driver);
+//        navigation = new Navigation(driver);
+//        privacy = new Privacy(driver);
+//    }
+//
+//    public AdminPanel getAdminPanel() {
+//        return adminPanel;
+//    }
+//
+//    public EventDetails getEventDetails() {
+//        return eventDetails;
+//    }
+//
+//    public Footer getFooter() {
+//        return footer;
+//    }
+//
+//    public Header getHeader() {
+//        return header;
+//    }
+//
+//    public LandingPage getLanding() {
+//        return landingPage;
+//    }
+//
+//    public Navigation getNavigation() {
+//        return navigation;
+//    }
+//
+//    public Privacy getPrivacy() {
+//        return privacy;
+//    }
+}
