+++ src/test/resources/app.properties
@@ -1,0 +1,3 @@
+baseURL = https://eventsexpress-test.azurewebsites.net/landing
+elementTimeOut = 10
+driver = chrome
